+++ how-to/workspace-platform-starter/client/types/module/shapes/app-shapes.d.ts
@@ -181,7 +181,7 @@ export interface ViewPreferenceUrl extends ViewPreference<PreferenceConstraintUr
 /**
  * A list of native related settings that can be updated.
  */
-export type NativePreferenceName = "args";
+export type NativePreferenceName = "arguments";
 /**
  * A list of web related settings that can be updated.
  */
@@ -213,7 +213,7 @@ export interface NativeLaunchOptions extends LaunchOptions {
 		/**
 		 * Arguments are set as an array for compatibility with appAssets, launchExternalProcess and Snap.
 		 */
-		args?: string[];
+		arguments?: string[];
 	};
 	/**
 	 * What can be specified when launching a native app. This is an array of named types to reflect the properties you are happy to be specified.